i32 tid

+++ litebox_common_linux/src/lib.rs
@@ -1076,10 +1076,7 @@ pub struct Task<Platform: litebox::platform::RawPointerProvider> {
     /// Parent Process ID
     pub ppid: i32,
     /// Thread ID
-    ///
-    /// Note tid is i32 on Linux but long on freebsd (though it always fits into i32).
-    /// Use i64 for consistency
-    pub tid: i64,
+    pub tid: i32,
     /// When a thread whose `clear_child_tid` is not `None` terminates, and it shares memory with other threads,
     /// the kernel writes 0 to the address specified by `clear_child_tid` and then executes:
     ///
